+++ pages/editBottle.php
@@ -13,74 +13,75 @@ while ($suppliers = mysqli_fetch_array($sup)){
 }
 ?>
 
-<style>
 
-.container {
-  max-width: 100%;
-  width: 100%;
-
-}
-</style>
 <div class="container">
-       <div class="text-center">
-          <div id="bottle_pic"><div class="loader"></div></div>
-       </div>
-       <div id="bottle-inf"></div>
-        <p>
-        Name: 
-          <input class="form-control" name="bottle-name" type="text"  id="bottle-name" value="<?=$bottle['name']?>" />
-        </p>
-        <p>            
-        Size (ml):
-          <input class="form-control" name="bottle-size" type="text" id="bottle-size" value="<?=$bottle['ml']?>"/>
-        </p>
-        <p>
-        Price:
-          <input class="form-control" name="bottle-price" type="text" id="bottle-price" value="<?=$bottle['price']?>"/>
-        </p>
- 		<p>
-        Height:
-          <input class="form-control" name="bottle-height" type="text" id="bottle-height" value="<?=$bottle['height']?>"/>
-        </p>
-        <p>
-        Width:
-          <input class="form-control" name="bottle-width" type="text" id="bottle-width" value="<?=$bottle['width']?>"/>
-        </p>
-        <p>
-        Diameter:
-          <input class="form-control" name="bottle-diameter" type="text" id="bottle-diameter" value="<?=$bottle['diameter']?>"/>
-        </p>
-        <p>
-        Stock (pieces):
-          <input class="form-control" name="bottle-pieces" type="text" id="bottle-pieces" value="<?=$bottle['pieces']?>"/>
-        </p>        
-        <p>
-        Notes:
-          <input class="form-control" name="bottle-notes" type="text" id="bottle-notes" value="<?=$bottle['notes']?>"/>
-        </p>
-        <p>
-        Supplier:
-          <select name="bottle-supplier" id="bottle-supplier" class="form-control">
-            <option value="" selected></option>
-            <?php foreach($supplier as $sup) { ?>
-				<option value="<?php echo $sup['name'];?>" <?php echo ($bottle['supplier']==$sup['name'])?"selected=\"selected\"":""; ?>><?php echo $sup['name'];?></option>
-		    <?php } ?>
-          </select>
-        </p>
-        <p>
-        Supplier URL:
-          <input class="form-control" name="bottle-supplier_link" type="text" id="bottle-supplier_link" value="<?=$bottle['supplier_link']?>"/>
-        </p>
-        <p>
-        Image:
-        <input type="file" name="bottle_pic_file" id="bottle_pic_file" class="form-control" />
-    	</p>            
+    <div class="text-center">
+        <div id="bottle_pic" class="mb-3">
+            <div class="loader"></div>
+        </div>
+    </div>
+    <div id="bottle-inf"></div>
+      <div class="row g-3">
+        <div class="col-md-4">
+            <label for="bottle-name" class="form-label">Name</label>
+            <input class="form-control" name="bottle-name" type="text" id="bottle-name" value="<?=$bottle['name']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-size" class="form-label">Size (ml)</label>
+            <input class="form-control" name="bottle-size" type="text" id="bottle-size" value="<?=$bottle['ml']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-price" class="form-label">Price:</label>
+            <input class="form-control" name="bottle-price" type="text" id="bottle-price" value="<?=$bottle['price']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-height" class="form-label">Height</label>
+            <input class="form-control" name="bottle-height" type="text" id="bottle-height" value="<?=$bottle['height']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-width" class="form-label">Width</label>
+            <input class="form-control" name="bottle-width" type="text" id="bottle-width" value="<?=$bottle['width']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-weight" class="form-label">Weight (grams)</label>
+            <input class="form-control" name="bottle-weight" type="text" id="bottle-weight" value="<?=$bottle['weight']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-diameter" class="form-label">Diameter</label>
+            <input class="form-control" name="bottle-diameter" type="text" id="bottle-diameter" value="<?=$bottle['diameter']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-pieces" class="form-label">Stock (pieces):</label>
+            <input class="form-control" name="bottle-pieces" type="text" id="bottle-pieces" value="<?=$bottle['pieces']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-notes" class="form-label">Notes</label>
+            <input class="form-control" name="bottle-notes" type="text" id="bottle-notes" value="<?=$bottle['notes']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-supplier" class="form-label">Supplier</label>
+            <select name="bottle-supplier" id="bottle-supplier" class="form-control">
+                <option value="" selected></option>
+                <?php foreach($supplier as $sup) { ?>
+                    <option value="<?php echo $sup['name'];?>" <?php echo ($bottle['supplier']==$sup['name'])?"selected=\"selected\"":""; ?>><?php echo $sup['name'];?></option>
+                <?php } ?>
+            </select>
+        </div>
+        <div class="col-md-4">
+            <label for="bottle-supplier_link" class="form-label">Supplier URL</label>
+            <input class="form-control" name="bottle-supplier_link" type="text" id="bottle-supplier_link" value="<?=$bottle['supplier_link']?>">
+        </div>
+        <div class="col-md-4">
+            <label for="bottle_pic_file" class="form-label">Image</label>
+            <input type="file" name="bottle_pic_file" id="bottle_pic_file" class="form-control">
+        </div>
         <div class="dropdown-divider"></div>
+        <div class="modal-footer">
+            <input type="submit" name="button" class="btn btn-primary" id="bottle-save" value="Save">
+        </div>
+    </div>
 </div>
-      <div class="modal-footer">
-        <input type="submit" name="button" class="btn btn-primary" id="bottle-save" value="Save">
-      </div>
-    </div>  
+ 
 <hr>
 <script>
 $('#bottle_pic').html('<img class="img-profile-avatar" src="<?=$doc['photo']?: '/img/logo_def.png'; ?>">');
@@ -99,6 +100,7 @@ $('#bottle-save').click(function() {
 			width: $("#bottle-width").val(),
 			diameter: $("#bottle-diameter").val(),
 			pieces: $("#bottle-pieces").val(),
+			weight: $("#bottle-weight").val(),
 			notes: $("#bottle-notes").val(),
 			supplier: $("#bottle-supplier").val(),
 			supplier_link: $("#bottle-supplier_link").val(),