+++ releasenotes.md
@@ -1,5 +1,7 @@
 Whats New in v10.0
 --------------------------
+- Bottle edit page format update
+- Weight added for the bottles inventory
 - Added Inventory for finished Compounds
 - Minor improvements in formula attachments section
 - For full details please refer to the CHANGELOG